Finishing Assigment
assignment 1 :
create 2 HTML files.

In the first HTML file, add the JS code in the HTML file.

In the second HTML file, add the JS code in a external JS file and reference the JS file in your HTML.

Write multiple messages to the console using console.log(). Each message should include at least a string, a number, and a boolean.